+++ tests/conftest.py
@@ -98,10 +98,13 @@ def connect_to_slave(pytestconfig, read_config):
 
 
 @pytest.fixture()
-def virtual_drive():
+def virtual_drive(read_config):
     test_ip = "127.0.0.1"
     test_port = 81
     server = VirtualDrive(test_ip, test_port)
     server.start()
-    yield server
+    net = EthernetNetwork()
+    protocol_contents = read_config["ethernet"]
+    virtual_servo = net.connect_to_slave(server.ip, protocol_contents["dictionary"], server.port)
+    yield server, virtual_servo
     server.stop()